+++ app/assets/javascripts/Components/Modals/submission_file_upload_modal.jsx
@@ -7,6 +7,7 @@ class SubmissionFileUploadModal extends React.Component {
     this.state = {
       newFiles: [],
       unzip: false,
+      renameTo: "",
     };
   }
 
@@ -16,7 +17,7 @@ class SubmissionFileUploadModal extends React.Component {
 
   onSubmit = event => {
     event.preventDefault();
-    this.props.onSubmit(this.state.newFiles, this.state.unzip);
+    this.props.onSubmit(this.state.newFiles, this.state.unzip, this.state.renameTo);
   };
 
   handleFileUpload = event => {
@@ -27,8 +28,80 @@ class SubmissionFileUploadModal extends React.Component {
     this.setState({unzip: !this.state.unzip});
   };
 
+  handleRenameChange = event => {
+    this.setState({renameTo: event.target.value});
+  };
+
+  fileRenameInputBox = () => {
+    let fileRenameInput;
+    let filesToShow;
+    if (this.props.uploadTarget) {
+      let filenames = this.props.requiredFiles.filter(
+        filename =>
+          filename.startsWith(this.props.uploadTarget) &&
+          !filename.includes("/", this.props.uploadTarget.length)
+      );
+      filesToShow = filenames.map(filename => filename.slice(this.props.uploadTarget.length));
+    } else {
+      filesToShow = this.props.requiredFiles.filter(filename => !filename.includes("/"));
+    }
+    if (this.props.onlyRequiredFiles) {
+      fileRenameInput = (
+        <select
+          className={"select-filename"}
+          onChange={this.handleRenameChange}
+          value={this.state.renameTo}
+          disabled={this.state.newFiles.length !== 1}
+          title={I18n.t("submissions.student.one_file_allowed")}
+          id={"rename-box"}
+        >
+          <option key={"select_file"}>{I18n.t("select_filename")}</option>
+          {filesToShow.map(filename => {
+            return (
+              <option key={filename} value={filename}>
+                {filename}
+              </option>
+            );
+          })}
+        </select>
+      );
+    } else if (this.props.requiredFiles.length >= 1) {
+      fileRenameInput = [
+        <datalist id="fileInput_datalist">
+          {filesToShow.map(filename => {
+            return <option key={filename} value={filename}></option>;
+          })}
+        </datalist>,
+        <input
+          className={"datalist-textbox"}
+          list="fileInput_datalist"
+          onChange={this.handleRenameChange}
+          placeholder={I18n.t("select_filename")}
+          value={this.state.renameTo}
+          disabled={this.state.newFiles.length !== 1}
+          title={I18n.t("submissions.student.one_file_allowed")}
+          id={"rename-box"}
+        />,
+      ];
+    } else {
+      fileRenameInput = (
+        <input
+          className={"file-rename-textbox"}
+          value={this.state.renameTo}
+          type={"text"}
+          name={"filename"}
+          onChange={this.handleRenameChange}
+          disabled={this.state.newFiles.length !== 1}
+          title={I18n.t("submissions.student.one_file_allowed")}
+          id={"rename-box"}
+        />
+      );
+    }
+    return fileRenameInput;
+  };
+
   clearState = () => {
-    this.setState({newFiles: []});
+    this.setState({newFiles: [], renameTo: ""});
   };
 
   render() {
@@ -62,6 +135,10 @@ class SubmissionFileUploadModal extends React.Component {
                 onChange={this.handleFileUpload}
               />
             </div>
+            <label htmlFor={"rename-box"}>
+              {I18n.t("submissions.student.rename_file_to")}&nbsp;
+              {this.fileRenameInputBox()}
+            </label>
             <div className={"modal-container"}>
               <input
                 type="submit"