+++ Changelog.md
@@ -41,9 +41,11 @@
 - Switch rendering of RMarkdown submission files to plaintext, rather than converting to HTML (#6008)
 - Ensure consistent order when displaying test results (#6010)
 - Added Capybara Gem (#6012)
+- Enable students to rename files during submission (#6045)
 - Separated peer review assignment upload/download from parent (#6048)
 - Expanded summary stats UI (#6050)
 - Fix bug for submission modal where students could submit files even if no file was chosen (#6052)
+- Removed searchbar used to filter files on submissions page (#6058)
 
 ## [v2.0.10]
 - Fix bug when sorting batch test runs where sorting by date was not working (#5906)